Component-ise chat sidebar, put in Layout, via Loader fn
