+++ src/components/navs/app-nav.tsx
@@ -42,7 +42,7 @@ const Nav = memo(function Nav({
 		<>
 			<div ref={ref}></div>
 			<div
-				className={`bg-background sticky border-b transition-colors ${!entry?.isIntersecting ? 'border-border' : 'border-transparent'} top-0 mb-4`}
+				className={`bg-background sticky border-b transition-colors ${!entry?.isIntersecting ? 'border-border' : 'border-transparent'} top-0 mb-2`}
 			>
 				<ScrollArea>
 					<NavigationMenu className="my-2">
